Changed default prefix to "d!"!
The original `n!` was only used during early stages of Dodo-Bot version 2. This also applies to beta builds of the bot before being publicly considered "stable".

This change is done to make the prefix match with the bot's project name and will not likely annoy you in any sort of way.